# docs/openapi/openapi.json
{
  "paths": {
...
    "/api/v1/students/{student_id}/results": {
      "get": {
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResultResponse"
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "tags": [
          "Students"
        ],
        "summary": "Get student results",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "student_id",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 0
          }
        }
      ]
    },
    "/api/v1/attendance/": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "description": "Page number (1-indexed)",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "required": false
          },
          {
            "in": "query",
            "name": "per_page",
            "description": "Results per page, max 100",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            },
            "required": false
          },
          {
            "in": "query",
            "name": "student_id",
            "schema": {
              "type": "integer"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "subject",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "start_date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "end_date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": false
          }
        ],
        "responses": {
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "tags": [
          "Attendance"
        ],
        "summary": "Get list of attendance records",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "DY Patil University ERP API",
    "version": "v1"
  },
...
  "openapi": "3.1.0",
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "Error code"
          },
          "status": {
            "type": "string",
            "description": "Error name"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "errors": {
            "type": "object",
            "description": "Errors"
          }
        }
      },
...
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "per_page": {
            "type": "integer"
          },
          "pages": {
            "type": "integer"
          },
          "has_next": {
            "type": "boolean"
          },
          "has_prev": {
            "type": "boolean"
          }
        }
      },
...
      "ResultResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "student_id": {
            "type": "integer"
          },
          "semester": {
            "type": "string"
          },
          "subject_id": {
            "type": "integer"
          },
          "internal_marks": {
            "type": "number"
          },
          "external_marks": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "grade": {
            "type": "string"
          },
          "is_published": {
            "type": "boolean"
          }
        }
      },
...
    },
    "responses": {
      "DEFAULT_ERROR": {
        "description": "Default error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "UNPROCESSABLE_ENTITY": {
        "description": "Unprocessable Entity",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}